requirements.txt
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
pandas>=2.2.0
selenium>=4.0.0
supabase>=2.0.0
ringcentral>=0.9.2
